rework export as aws sam app
