--- data-serving/scripts/data-pipeline/requirements.txt
@@ -1,1 +1,0 @@
-../convert-data/requirements.txt
