+++ data-serving/scripts/export-data/.aws-sam/build.toml
@@ -1,0 +1,32 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions]
+[function_build_definitions.92c1f9c3-c322-4006-8590-31aea86ca028]
+codeuri = "functions/01-split/"
+runtime = "python3.8"
+source_md5 = ""
+packagetype = "Zip"
+functions = ["DataExportSplitFunction"]
+
+[function_build_definitions.dcee1ecd-ad3f-4220-b46f-89ae6a87178e]
+codeuri = "functions/02-export/"
+runtime = "python3.8"
+source_md5 = ""
+packagetype = "Zip"
+functions = ["DataExportExportFunction"]
+
+[function_build_definitions.65b6c970-39a5-4fd1-bf6a-392894e81b0d]
+codeuri = "functions/03-parse/"
+runtime = "python3.8"
+source_md5 = ""
+packagetype = "Zip"
+functions = ["DataExportParseFunction"]
+
+[function_build_definitions.804fbaa7-29b8-419e-943b-98908d25971c]
+codeuri = "functions/04-combine/"
+runtime = "python3.8"
+source_md5 = ""
+packagetype = "Zip"
+functions = ["DataExportCombineFunction"]
+
+[layer_build_definitions]
