--- data-serving/scripts/export-data/01-export-and-split/case_fields.yaml
@@ -1,117 +1,0 @@
-- name: _id
-  description: Unique ID of the case
-- name: caseReference.verificationStatus
-  description: >
-    Whether this case was verified by a curator.
-    This is verification at the ingestion level, that is curators verified that
-    the data that got ingested corresponds to the data that was present in an
-    source outside the system.
-    It does not mean that the curators verified that the case really existed.
-    Values can be VERIFIED, UNVERIFIED or EXCLUDED. EXCLUDED cases contain a note
-    describing the reason for exclusion and when it was done.
-- name: caseReference.sourceId
-  description: Unique ID of the source for this case
-- name: caseReference.sourceUrl
-  description: URL of the raw source for this case
-- name: caseReference.sourceEntryId
-  description: The id of this case that was assigned by the source
-- name: caseReference.additionalSources
-  description: Additional sources that provided data for this case
-- name: demographics.ageRange.start
-  description: Minimum age bracket, included
-- name: demographics.ageRange.end
-  description: Minimum age bracket, excluded if different than start age
-- name: demographics.gender
-  description: Male, Female, Non-binary/Third gender, Other
-- name: demographics.occupation
-  description: Primary occupation of the case if any
-- name: demographics.nationalities
-  description: Nationalities of the case in English.
-- name: demographics.ethnicity
-  description: >
-    Ethnicity of the case, Asian, Caucasian, etc. Social group that has a common
-    national or cultural tradition.
-- name: genomeSequences
-  description: Genome sequences of the pathogen
-- name: location.country
-  description: English name of country this case was reported in
-- name: location.administrativeAreaLevel1
-  description: First administrative subdivision of a country
-- name: location.administrativeAreaLevel2
-  description: Second administrative subdivision of a country
-- name: location.administrativeAreaLevel3
-  description: Third administrative subdivision of a country
-- name: location.geoResolution
-  description: How granular the location is, i.e. Country or Admin2
-- name: location.name
-  description: >
-    Full name of the location, i.e. "Lyon, Auvergne-Rhône-Alpes, France"
-- name: location.geometry.latitude
-  description: Latitude of the centroid of the location
-- name: location.geometry.longitude
-  description: Longitude of the centroid of the location
-- name: events.confirmed.value
-  description: Confirmation method
-- name: events.confirmed.date
-  description: Confirmed case date
-- name: events.onsetSymptoms.date
-  description: Symptom onset date
-- name: events.firstClinicalConsultation.date
-  description: First clinical consultation
-- name: events.selfIsolation.date
-  description: Date of self isolation
-- name: events.hospitalAdmission.value
-  description: Hospital admission
-- name: events.hospitalAdmission.date
-  description: Hospital admission date
-- name: events.icuAdmission.value
-  description: Admission to isolation unit
-- name: events.icuAdmission.date
-  description: Date admitted to isolation unit
-- name: events.outcome.value
-  description: Outcome
-- name: events.outcome.date
-  description: Outcome date
-- name: symptoms.status
-  description: >
-    Symptom status if known: Asymptomatic, Symptomatic or Presymptomatic
-- name: symptoms.values
-  description: List of symptoms
-- name: preexistingConditions.hasPreexistingConditions
-  description: Whether the case had known preexisting conditions
-- name: preexistingConditions.values
-  description: Preexisting conditions if any
-- name: transmission.routes
-  description: List of routes of transmission
-- name: transmission.places
-  description: List of places of transmission
-- name: transmission.linkedCaseIds
-  description: List of case ids for which transmission occurred with this case
-- name: travelHistory.traveledPrior30Days
-  description: >
-    Whether the case travelled in the 30 days prior to becoming infected
-- name: travelHistory.travel
-  description: Array of locations the case travelled to
-- name: pathogens
-  description: List of pathogens for this case
-- name: notes
-  description: >
-    Anything that does not fit in a schema but worth mentioning about the case
-- name: revisionMetadata.revisionNumber
-  description: Revision number (smaller is older)
-- name: revisionMetadata.creationMetadata.curator
-  description: Email of the curator who provided this case
-- name: revisionMetadata.creationMetadata.date
-  description: Date this case was first entered in the system
-- name: revisionMetadata.creationMetadata.notes
-  description: Creation notes from curator
-- name: revisionMetadata.updateMetadata.curator
-  description: Email of the curator who edited the case
-- name: revisionMetadata.updateMetadata.date
-  description: Date an update was made to the case
-- name: revisionMetadata.updateMetadata.notes
-  description: Update notes from curator
-- name: exclusionData.note
-  description: Note that was entered by curator when marking the case as excluded
-- name: exclusionData.date
-  description: Date this case was marked as excluded
