--- data-serving/scripts/export-data/01-export-and-split/requirements.txt
@@ -1,1 +1,0 @@
-pyyaml==5.3.1
