+++ data-serving/scripts/export-data/functions/01-split/requirements.txt
@@ -1,0 +1,3 @@
+boto3==1.16.60
+pymongo==3.11.2
+dnspython==2.1.0
