+++ data-serving/scripts/export-data/functions/02-export/fields.txt
@@ -1,0 +1,47 @@
+_id
+caseReference.additionalSources
+caseReference.sourceEntryId
+caseReference.sourceId
+caseReference.sourceUrl
+caseReference.uploadIds
+caseReference.verificationStatus
+demographics.ageRange.end
+demographics.ageRange.start
+demographics.ethnicity
+demographics.gender
+demographics.nationalities
+demographics.occupation
+events
+genomeSequences
+location.administrativeAreaLevel1
+location.administrativeAreaLevel2
+location.administrativeAreaLevel3
+location.country
+location.geoResolution
+location.geometry.latitude
+location.geometry.longitude
+location.name
+location.place
+location.query
+notes
+pathogens
+preexistingConditions.hasPreexistingConditions
+preexistingConditions.values
+revisionMetadata.creationMetadata.curator
+revisionMetadata.creationMetadata.date
+revisionMetadata.creationMetadata.notes
+revisionMetadata.editMetadata.curator
+revisionMetadata.editMetadata.date
+revisionMetadata.editMetadata.notes
+revisionMetadata.revisionNumber
+symptoms.status
+symptoms.values
+transmission.linkedCaseIds
+transmission.places
+transmission.routes
+travelHistory.travel.dateRange.end
+travelHistory.travel.dateRange.start
+travelHistory.travel.location.name
+travelHistory.travel.methods
+travelHistory.travel.purpose
+travelHistory.traveledPrior30Days
