+++ data-serving/scripts/export-data/functions/02-export/requirements.txt
@@ -1,0 +1,1 @@
+boto3==1.16.60
