+++ data-serving/scripts/export-data/functions/04-combine/requirements.txt
@@ -1,0 +1,1 @@
+boto3==1.16.60
