+++ data-serving/scripts/export-data/samconfig.toml
@@ -1,0 +1,10 @@
+version = 0.1
+[default]
+[default.deploy]
+[default.deploy.parameters]
+stack_name = "data-export"
+s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-1txyh416iht6x"
+s3_prefix = "data-export"
+region = "us-east-1"
+confirm_changeset = true
+capabilities = "CAPABILITY_IAM"
