+++ data-serving/scripts/export-data/tests/requirements.txt
@@ -1,0 +1,3 @@
+pytest
+pytest-mock
+boto3
