+++ data-serving/scripts/export-data/tests/unit/test_checker.py
@@ -1,0 +1,6 @@
+from functions.stock_checker import app
+
+
+def test_stock_checker():
+    data = app.lambda_handler(None, "")
+    assert 0 <= data["stock_price"] > 0 <= 100
