+++ data-serving/scripts/export-data/tests/unit/test_seller.py
@@ -1,0 +1,17 @@
+from functions.stock_seller import app
+
+
+def test_stock_checker():
+    stock_price = 25
+    input_payload = {"stock_price": stock_price}
+
+    data = app.lambda_handler(input_payload, "")
+
+    assert "id" in data
+    assert "price" in data
+    assert "type" in data
+    assert "timestamp" in data
+    assert "qty" in data
+
+    assert data["type"] == "sell"
+    assert data["price"] == str(stock_price)
